Functionality and style modifications
Edited the search input styling - made it look more or less like the other buttons so they'll generally look the same.
Added some documentation.
Functionally made the button to not interfere with the flow of the program : 1) When a user clicks out of bounds of the input, the search box won't be focused, thus the user won't type in it mistakenly.
2) Keys which are binded to actions (like R=rotate), won't be activated when searching in search bar.